Pretty massive overhaul of the skeleton code
I had to move a lot of things around to make them work well with the
twitter build script and the appengine uploader.  I also have built
the first three (static, currently) pages of this website.

Now, all static files are in view/static; templates are in view/templates;
and the static link pointing to the current version of the directory I
want to upload is named 'frontend', and points to the 'view' directory
on the dev server.  The build script has been updated accordingly.

This is really just a first draft.  Much more work needs to be done,
including getting a favicon set up, moving the about and lore sections
into a database, making a database-driven user and members section,
forums, probably quite a bit of design work, the logo could use more
detail...

But I'm pretty happy with this as a start.