Initial admin pages
I created the code to handle database loading of data.  Right now,
only the Lore section is done, but the important part was getting
the first touches of database logic written.  There is also an
admin page that can be used to create, edit, and delete the Lore
sections.

Next, I'll be refactoring what I just did to make it generic enough
that I can just extend it, renaming it all with "About" instead of
"Lore" and have all the rest of the logic work the same.