Added an admin UI for setting member ranks.
Many of the changes in this relate to minor changes in how it is determined
whether or not a user has site administration permissions.  This should make
it easier in the future to manage those permissions.

Many more of the changes relate to the account preferences form, specifically
for the admin page.  I wanted to update it so that the form would automatically
limit available ranks in the dropdown to only the current user's rank and lower.
It appears to work, but the UI for non-site-admins to change that is not done
yet.

Anything else is related to minor bug fixes, mainly revolving around a slight
change in the structure of the admin pages.